+++ src/agdb/collections.rs
@@ -1,10 +1,6 @@
 pub mod bit_set;
 pub mod dictionary;
-pub mod dictionary_storage;
 pub mod indexed_map;
-pub mod map_storage;
+pub mod map;
 pub mod multi_map;
-pub mod multi_map_storage;
 pub mod vec;
-
-mod map;
