+++ src/agdb/command.rs
@@ -1,4 +1,4 @@
-use crate::collections::dictionary::dictionary_index::DictionaryIndex;
+use crate::collections::dictionary::DictionaryIndex;
 use crate::db::db_key_value_index::DbKeyValueIndex;
 use crate::graph::GraphIndex;
 use crate::DbId;
